requirements.txt
asgiref==3.4.1
Django==3.2.5
django-image-uploader-widget==0.0.3
mysqlclient==2.0.3
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
